Capture the orderly and exception registers
On a HW exception, we can look back on the stack and find the preserved
state of the system when the crash happened, so use that.

On an orderly exception (C++ excetion, abort, panic, etc.), we catch the
state of the machine ASAP in the highest-level call and merge it with
state captured later, to better ensure GDB can get a backtrace.