Update ESP8266WifiGeneric.cpp to allow compile time selection of DNS resolution precedence IPv4/IPv6
eg. platformio uses prebuilt liblwip6-1460-feat.a which is built with DNS resolution order IPV4 before IPv6. This causes problems in IPv6 only environments if there is a DNS record with A and AAAA. lwip then tries to connect to the resolved v4 even if there is no global IPv4 address. Didn't find a way to get pio rebuild the complete library. That's why I patched ESP8266WifiGeneric.cpp to use dns_gethostbyname_addrtype instead of dns_gethostbyname (which is only a wrapper for gethostbyname_addrtype and the define to prefer v4). With this change it's possible to select resolution beaviour during compile time.