Fix integer overflow in basic password authentication
The string lengths were added and then stored in "char" which is
limited in values at most up to 255. Using size_t is a much more
appropriate type.

In addition the code was using base64 with newlines injected
(easy to fall into that trap as the default is imho wrong), which
means that anything longer than ~ 60 characters never matched and
you had no way to authenticate against the server.