Allow byte access to PROGMEM without pgm_read macro
This PR pulls in code from @pvvx's esp8266web server at
https://github.com/pvvx/esp8266web (licensed in the public domain)
to allow reads of 8 and 16-bit values out of flash without the need
for pgm_read_byte/word macros.  It installs an exception handler that
decodes the faulting instruction and simulates is properly and returns
immediately to user code.  That said, it is significantly slower (20x?)
than using the pgm_read macros properly.