HTTPClient: implement Stream class
It is now possible to stream HTTP request data without knowing in
advance the data size (using chunked transfer encoding), and to
stream HTTP response data (correctly decoded according to the
transfer encoding method used by the server) without requiring an
output Stream implementation.
(The existing getStream() method in the HTTPClient class returns
the underlying WiFiClient stream which is not aware of the HTTP
transfer encoding.)
When using chunked encoding in an HTTP request, each call to
write() results in the transmission of a separate chunk.

The setting of _transferEncoding to HTTPC_TE_IDENTITY in the
handleHeaderResponse() method has been removed because
_transferEncoding is now used for both transmission and reception,
and is reset to HTTPC_TE_IDENTITY in the clear() method.