SPI: Add async write functions
Add new SPI write functions suffixed with
underscore that are allowed to return
before SPI data is fully written.

Internally, this allows the CPU to run up to
one SPI FIFO ahead of the SPI transmitter.

A benchmark: Updating a string of 30 APA102
LEDs on 160MHz ESP8266:
    235us - FastGPIO bitbang
	 97us - SPI.write32() per pixel
	 79us - SPI.write32_() per pixel

This is particularly useful for complex pixel
calculations. Approximately 1.5 microseconds
of calculations per pixel can now be pipelined
in parallel with the previous pixel transmission.