Move waveform vars to a struct to save IRAM
When globals are used, GCC needs to load an address from literal memory
and then load/store the value.  Even if the two globals are immediately
following each other in the code, GCC can't know this at compile time so
emits the indirect sequence.

Move the globals to a struct which allows GCC to load the address in a
single variable and then use base+offset addressing to access related
globals.  This reduces code size and increases code speed somewhat.

IRAM usage for the ISR went from 0x263 to 0x231 bytes, a savings of 50
bytes.