Favor phase over duty cycle
As mentioned by @devyte and others, when you calculate the time for a
new edge, if you use the exact time you're doing it instead of having
edges = n * period, you will have slip due to there being some delta
between the ideal (n*period) and sum(nowX, period).  That could cause
slippage of edges in absolute terms.

Simply assume the edge was handled on the exact time it wanted it to be
handled.  This may trade off duty cycle accuracy for edge accuracy, but
it seems that the edge timing is more critical in most apps.  In steady
state, the delta between now(n) and (n*period) should be close to
constant and therefore cancel itself out.